# docs/api-documentation.yml
openapi: 3.0.3
info:
  title: UberFoods API
  version: 1.0.0
  description: Complete API documentation for UberFoods food delivery platform
  contact:
    name: UberFoods API Support
    email: api@uberfoods.com
    url: https://uberfoods.com/api-support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.uberfoods.com/v1
    description: Production server
  - url: https://staging-api.uberfoods.com/v1
    description: Staging server
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /auth/register:
    post:
      tags: ["Authentication"]
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
                - role
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "securepassword123"
                firstName:
                  type: string
                  example: "John"
                lastName:
                  type: string
                  example: "Doe"
                role:
                  type: string
                  enum: ["customer", "driver", "restaurant_owner"]
                  example: "customer"
                phone:
                  type: string
                  example: "+1234567890"
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken:
                    type: string
                    example: "refresh_token_here"
        400:
          $ref: '#/components/responses/BadRequest'
        409:
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: ["Authentication"]
      summary: User login
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                  refreshToken:
                    type: string
        401:
          $ref: '#/components/responses/Unauthorized'

  # Restaurants
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "user_123"
        email:
          type: string
          format: email
          example: "user@example.com"
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        role:
          type: string
          enum: ["customer", "driver", "restaurant_owner"]
          example: "customer"
        phone:
          type: string
          example: "+1234567890"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Restaurant schemas
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "VALIDATION_ERROR"
              message:
                type: string
                example: "Invalid input data"
              details:
                type: object
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "UNAUTHORIZED"
              message:
                type: string
                example: "Invalid or missing authentication token"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "CONFLICT"
              message:
                type: string
                example: "Resource already exists or conflict detected"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for server-to-server communication